Move all the (de)serialisation methods to separate file.
